Split ProcessedDefinitionsValidator into sub-validators
ProcessedDefinitionsValidator is big and hard to navigate.
This extracts its validation methods into focused
validator classes, each responsible for a specific domain,
while keeping ProcessedDefinitionsValidator as a thin orchestrator.
