+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/validation/ProcessedDefinitionsValidator.java
@@ -1,2212 +1,64 @@
 package no.sikt.graphitron.validation;
 
-import no.sikt.graphitron.configuration.GeneratorConfig;
-import no.sikt.graphitron.definitions.fields.*;
-import no.sikt.graphitron.definitions.fields.containedtypes.FieldReference;
-import no.sikt.graphitron.definitions.fields.containedtypes.MutationType;
-import no.sikt.graphitron.definitions.helpers.CodeReferenceWrapper;
-import no.sikt.graphitron.definitions.helpers.ServiceWrapper;
-import no.sikt.graphitron.definitions.interfaces.FieldSpecification;
-import no.sikt.graphitron.definitions.interfaces.GenerationField;
-import no.sikt.graphitron.definitions.interfaces.GenerationTarget;
-import no.sikt.graphitron.definitions.interfaces.RecordObjectSpecification;
-import no.sikt.graphitron.definitions.mapping.JOOQMapping;
-import no.sikt.graphitron.definitions.objects.*;
-import no.sikt.graphitron.definitions.sql.SQLCondition;
-import no.sikt.graphitron.generators.codebuilding.VariablePrefix;
-import no.sikt.graphitron.generators.context.InputParser;
-import no.sikt.graphitron.mappings.ReflectionHelpers;
-import no.sikt.graphitron.mappings.TableReflection;
-import no.sikt.graphitron.validation.messages.InputTableError;
-import no.sikt.graphql.directives.GenerationDirectiveParam;
-import no.sikt.graphql.naming.GraphQLReservedName;
+import no.sikt.graphitron.definitions.fields.ObjectField;
 import no.sikt.graphql.schema.ProcessedSchema;
-import org.apache.commons.text.similarity.LevenshteinDistance;
-import org.jooq.Field;
-import org.jooq.Key;
 
-import java.lang.reflect.Method;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
-import java.util.*;
-import java.util.function.Function;
+import java.util.Collections;
+import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import static java.util.stream.Collectors.groupingBy;
-import static no.sikt.graphitron.configuration.GeneratorConfig.shouldMakeNodeStrategy;
-import static no.sikt.graphitron.configuration.Recursion.recursionCheck;
-import static no.sikt.graphitron.generators.context.NodeIdReferenceHelpers.*;
-import static no.sikt.graphitron.mappings.JavaPoetClassName.STRING;
-import static no.sikt.graphitron.mappings.ReflectionHelpers.getJooqRecordClassForNodeIdInputField;
-import static no.sikt.graphitron.mappings.TableReflection.*;
-import static no.sikt.graphitron.validation.ServiceMethodFormatter.formatExpectedListableType;
-import static no.sikt.graphitron.validation.ServiceMethodFormatter.formatOverloads;
-import static no.sikt.graphitron.validation.ValidationHandler.*;
-import static no.sikt.graphitron.validation.messages.ConstructError.*;
-import static no.sikt.graphitron.validation.messages.InputTableError.MISSING_FIELD;
-import static no.sikt.graphitron.validation.messages.InputTableError.MISSING_NON_NULLABLE;
-import static no.sikt.graphitron.validation.messages.MultitableError.MISSING_TABLE_ON_MULTITABLE;
-import static no.sikt.graphql.directives.GenerationDirective.*;
-import static no.sikt.graphql.directives.GenerationDirective.NODE_ID;
-import static no.sikt.graphql.directives.GenerationDirectiveParam.TYPE_ID;
-import static no.sikt.graphql.naming.GraphQLReservedName.*;
-import static org.apache.commons.lang3.StringUtils.*;
+import static no.sikt.graphitron.validation.ValidationHandler.logWarnings;
+import static no.sikt.graphitron.validation.ValidationHandler.throwIfErrors;
 
 /**
- * Class for producing warnings related to potential issues in the defined schema.
- * This is only used before running generation, but generally does not prohibit further execution.
- * The intention is that the warnings should provide information on potential issues should an issue occur later.
+ * Orchestrates schema validation by delegating to focused sub-validators.
+ * Each sub-validator is responsible for a specific domain of validation.
  */
 public class ProcessedDefinitionsValidator {
-    protected final ProcessedSchema schema;
+    private final ProcessedSchema schema;
     private final List<ObjectField> allFields;
+
+    private final List<AbstractSchemaValidator> validators;
 
     public ProcessedDefinitionsValidator(ProcessedSchema schema) {
         this.schema = schema;
-        allFields = schema
+        allFields = Collections.unmodifiableList(schema
                 .getObjects()
                 .values()
                 .stream()
                 .flatMap(it -> it.getFields().stream())
-                .collect(Collectors.toList());
+                .collect(Collectors.toList()));
+
+        validators = List.of(
+                new InterfaceValidator(schema, allFields),
+                new TableValidator(schema, allFields),
+                new RecordMappingValidator(schema, allFields),
+                new ServiceValidator(schema, allFields),
+                new MutationValidator(schema, allFields),
+                new FieldDirectiveValidator(schema, allFields),
+                new NodeValidator(schema, allFields),
+                new PaginationValidator(schema, allFields),
+                new FederationValidator(schema, allFields)
+        );
     }
 
     /**
      * Validate the directive usage in the schema.
      */
     public void validateDirectiveUsage() {
-        schema.getObjects().values().forEach(it -> checkPaginationSpecs(it.getFields()));
-
         // Validate cycles first to ensure there are no infinite loops later
-        validateNoCycles();
+        new CycleValidator(schema, allFields).validate();
+        throwIfErrors();
 
-        validateMultitableTypeTables();
-        validateTablesAndKeys();
-        validateRequiredTableFields();
-        validateSingleTableInterfaceDefinitions();
-        validateInterfacesReturnedInFields();
-        validateMultitableFieldsOutsideRoot();
-        validateTypesUsingNodeInterfaceWithoutNodeDirective();
-        validateInputFields();
-        validateExternalMappingReferences();
-        validateServiceMethods();
-        validateServiceInputTypes();
-        validateServiceMethodParameterTypes();
-        validateMutationDirectives();
-        validateMutationRequiredFields();
-        validateRecursiveRecordInputs();
-        validateOnlyOneInputRecordInputWhenNoTypeTableIsPresent();
-        validateNotUsingBothExternalFieldAndField();
-        validateConstructTypeDirective();
-        validateExternalField();
-        validateNodeDirective();
-        validateUnionAndInterfaceSubTypes();
-        validateNodeId();
-        validateNodeIdReferenceInJooqRecordInput();
-        validateNodeIdFieldsInJavaRecordInputs();
-        validateSplitQueryFieldsInJavaRecords();
-        validateWrapperTypesWithPreviousTable();
-        validateJavaRecordFieldMappings();
-        validatePaginatedFieldsHaveOrdering();
-        validateOrderByInputTypes();
-        validateDefaultOrderNotOnInterfaceOrUnion();
-        validateLookupArguments();
-        validateEntities();
-
+        validators.forEach(AbstractSchemaValidator::validate);
         logWarnings();
         throwIfErrors();
     }
 
-    /*
-     * This is a temporary validation until GG-104 has been fixed.
-     */
-    private void validateUnionAndInterfaceSubTypes() {
-        schema.getObjects()
-                .values().stream()
-                .flatMap(o -> o.getFields().stream())
-                .filter(schema::isMultiTableField)
-                .filter(field -> !field.getTypeName().equals(NODE_TYPE.getName()))
-                .filter(field -> !field.getTypeName().equals(FEDERATION_SERVICE_TYPE.getName()))
-                .filter(field -> !field.getTypeName().equals(FEDERATION_ENTITY_UNION.getName()))
-                .forEach(field -> {
-                    var subTypes = schema.getTypesFromInterfaceOrUnion(field.getTypeName()).orElse(List.of());
-                    if (subTypes.size() < 2 && subTypes.stream().noneMatch(type -> type.implementsInterface(ERROR_TYPE.getName()))) {
-                        addErrorMessage(
-                                "Multitable queries is currently only supported for interface and unions with more than one implementing type. \n" +
-                                        "The field %s's type %s has %d implementing type(s).", field.getName(), field.getTypeName(), subTypes.size());
-                    }
-                });
-    }
-
-    private void validateNodeDirective() {
-        var typeIdMap = new HashMap<String, Set<String>>();
-        schema.getObjects().values().stream()
-                .filter(ObjectDefinition::hasNodeDirective)
-                .forEach(objectDefinition -> {
-                    var nodeConfiguration = objectDefinition.getNodeConfiguration()
-                            .orElseThrow(() -> new IllegalStateException(String.format( "Node type '%s' is unexpectedly missing node configuration.", objectDefinition.getName())));
-
-                    if (objectDefinition.getTable() == null) {
-                        addErrorMessage("Type %s has the %s directive, but is missing the %s directive.",
-                                objectDefinition.getName(), NODE.getName(), TABLE.getName());
-                    } else {
-                        if (!objectDefinition.getTable().equals(nodeConfiguration.targetTable())) {
-                            throw new IllegalStateException(String.format(
-                                    "Unexpected mismatch between node configuration table and object table for '%s'. Please report if you have encountered this issue.",
-                                    objectDefinition.getName())
-                            );
-                        }
-
-                        var tableFields = getJavaFieldNamesForTable(nodeConfiguration.targetTable().getName());
-                        nodeConfiguration.keyColumnsJavaNames().stream()
-                                .filter(col -> tableFields.stream().noneMatch(it -> it.equalsIgnoreCase(col)))
-                                .forEach(col -> addErrorMessage(
-                                        "Key column '%s' in node ID for type '%s' does not exist in table '%s'",
-                                        col,
-                                        objectDefinition.getName(),
-                                        nodeConfiguration.targetTable().getName())
-                                );
-
-                        if (getPrimaryOrUniqueKeyMatchingFields(nodeConfiguration.targetTable().getName(), nodeConfiguration.keyColumnsJavaNames()).isEmpty()) {
-                            addErrorMessage(
-                                    "Key columns in node ID for type '%s' does not match a PK/UK for table '%s'",
-                                    objectDefinition.getName(),
-                                    nodeConfiguration.targetTable().getName());
-                        }
-                    }
-                    if (!objectDefinition.implementsInterface(NODE_TYPE.getName())) {
-                        addErrorMessage("Type %s has the %s directive, but does not implement the %s interface.",
-                                objectDefinition.getName(), NODE.getName(), NODE_TYPE.getName());
-                    }
-                    if (GeneratorConfig.requireTypeIdOnNode() && !nodeConfiguration.hasCustomTypeId()) {
-                        addErrorMessage("Type '%s' has the '%s' directive, but is missing the '%s' parameter which has been configured to be required.",
-                                objectDefinition.getName(), NODE.getName(), TYPE_ID.getName());
-                    }
-                    typeIdMap.computeIfAbsent(nodeConfiguration.typeId(), k -> new HashSet<>())
-                            .add(objectDefinition.getName());
-                });
-
-        typeIdMap.entrySet().stream()
-                .filter(it -> it.getValue().size() > 1)
-                .forEach((entry) ->
-                        addErrorMessage(
-                                "Multiple node types (%s) have the same node type ID '%s'. Type IDs must be unique.",
-                                String.join(", ", entry.getValue()),
-                                entry.getKey())
-                );
-    }
-
-    private void validateNodeId() {
-        Stream<? extends GenerationSourceField<?>> inputStream =
-                Stream.concat(
-                        schema.getInputTypes().values().stream().flatMap(it -> it.getFields().stream()),
-                        allFields.stream().flatMap(it -> it.getNonReservedArguments().stream())
-                );
-        Stream.concat(allFields.stream(), inputStream)
-                .filter(GenerationSourceField::hasNodeID)
-                .forEach(this::checkNodeId);
-    }
-
-    private void checkNodeId(GenerationField field) {
-        var fieldName = field instanceof ArgumentField
-                ? String.format("argument '%s' on a field in type '%s'", field.getName(), field.getContainerTypeName())
-                : String.format("field %s", field.formatPath());
-
-        if (!(field.isID() || field.getTypeName().equals(STRING.className.simpleName()))) {
-            addErrorMessage(
-                    "%s has %s directive, but is not an ID or String field.",
-                    capitalize(fieldName),
-                    NODE_ID.getName()
-            );
-        }
-
-        if (field.getNodeIdTypeName().isPresent()) {
-            var providedType = schema.getObject(field.getNodeIdTypeName().get());
-            if (providedType == null) {
-                addErrorMessage(
-                        "Type with name '%s' referenced in the %s directive for %s does not exist.",
-                        field.getNodeIdTypeName().get(),
-                        NODE_ID.getName(),
-                        fieldName
-                );
-                return;
-            } else if (!providedType.hasNodeDirective()) {
-                addErrorMessage(
-                        "Referenced type '%s' referenced in the %s directive for %s is missing the necessary %s directive.",
-                        providedType.getName(),
-                        NODE_ID.getName(),
-                        fieldName,
-                        NODE.getName()
-                );
-                return;
-            }
-        } else if (schema.getNodeTypeForNodeIdField(field).isEmpty()) { // Implicit nodeType
-            addErrorMessage("Cannot automatically deduce node type for node ID field %s. " +
-                            "Please specify the node type with the %s parameter in the %s directive.",
-                    field.formatPath(),
-                    GenerationDirectiveParam.TYPE_NAME.getName(),
-                    NODE_ID.getName()
-            );
-            return;
-        }
-
-        if (field instanceof ObjectField && isNodeIdReferenceField((ObjectField) field, schema)) {
-            // Only filter object fields because we currently don't have reference validation on input (GGG-209)
-            var recordType = Optional
-                    .ofNullable(schema.getRecordType(field.getContainerTypeName()))
-                    .flatMap(it -> Optional.ofNullable(it.getTable()));
-
-            var referenceTable = schema.getNodeTypeForNodeIdFieldOrThrow(field) // This should have been validated above
-                    .getTable();
-            recordType.ifPresent(it -> validateReferencePath(field, it.getMappingName(), referenceTable.getMappingName()));
-            if (recordType.isEmpty()) {
-                var inputMapping = schema.findInputTables(field).stream().findFirst();
-                if (inputMapping.isPresent() && !inputMapping.get().equals(referenceTable)) {
-                    validateReferencePath(field, inputMapping.get().getMappingName(), referenceTable.getMappingName());
-                }
-            }
-        }
-
-        if (field.hasFieldDirective()) {
-            // Allow @nodeId + @field only in @record input types targeting jOOQ record fields
-            boolean isAllowed = getJooqRecordClassForNodeIdInputField(field, schema).isPresent();
-
-            if (!isAllowed) {
-                addErrorMessage(
-                        "%s has both the '%s' and '%s' directives, which is only supported for node ID fields in Java Record inputs.",
-                        capitalize(fieldName),
-                        NODE_ID.getName(),
-                        FIELD.getName()
-                );
-            }
-        }
-        if (field.isExternalField()) {
-            addErrorMessage(
-                    "%s has both the '%s' and '%s' directives, which is not supported.",
-                    capitalize(fieldName),
-                    NODE_ID.getName(),
-                    EXTERNAL_FIELD.getName()
-            );
-        }
-    }
-
-    private void validateNodeIdReferenceInJooqRecordInput() {
-        schema.getInputTypes()
-                .values()
-                .stream()
-                .filter(it -> schema.isRecordType(it.getName()))
-                .map(it -> schema.getRecordType(it.getName()))
-                .filter(it -> it.getFields().stream().anyMatch(FieldSpecification::hasNodeID))
-                .filter(it -> it.hasTable() && getTableByJavaFieldName(it.getTable().getName()).isPresent()) // No need to continue validation here if table does not exist. Table names is validated in validateTablesAndKeys
-                .forEach(jooqRecordInput ->
-                        jooqRecordInput.getFields()
-                                .stream()
-                                .filter(schema::isNodeIdField)
-                                .filter(it -> schema.getNodeTypeForNodeIdField(it).isPresent()) // This is validated in checkNodeId
-                                .filter(it -> !schema.getNodeTypeForNodeIdFieldOrThrow(it).getTable().equals(jooqRecordInput.getTable()))
-                                .forEach(it -> {
-                                    validateNodeIdReferenceInRecord(jooqRecordInput.getTable().getName(), it, true, jooqRecordInput.getTable());
-                                    findForeignKeyForNodeIdField(it, schema, jooqRecordInput.getTable()).ifPresent(foreignKey -> {
-                                        if (isUsedInUpdateMutation(jooqRecordInput)) {
-                                            getPrimaryKeyForTable(jooqRecordInput.getTable().getName())
-                                                    .map(Key::getFields)
-                                                    .filter(pkFields -> pkFields.stream().anyMatch(pkF -> foreignKey.getFields().stream().anyMatch(pkF::equals)))
-                                                    .stream().findFirst()
-                                                    .ifPresent((a) -> addErrorMessage(
-                                                            "Foreign key used for node ID field '%s' in jOOQ record input '%s' overlaps with the primary key of the jOOQ record table. This is not supported for update/upsert mutations.",
-                                                            it.getName(),
-                                                            it.getContainerTypeName()
-                                                    ));
-                                        }
-                                    });
-                                })
-                );
-    }
-
-    private void validateNodeIdReferenceInRecord(String jooqRecordName, GenerationField field, boolean isJooqRecordInput, JOOQMapping currentTable) {
-        var inputKind = isJooqRecordInput ? "jOOQ record input" : "Java record input";
-
-        var foreignKeyOptional = findForeignKeyForNodeIdField(field, schema, currentTable);
-        if (foreignKeyOptional.isEmpty()) {
-            addErrorMessage("Cannot find foreign key for node ID field '%s' in %s '%s'.",
-                    field.getName(),
-                    inputKind,
-                    field.getContainerTypeName());
-            return;
-        }
-
-        var foreignKey = foreignKeyOptional.get();
-
-        if (!foreignKey.getTable().getName().equalsIgnoreCase(jooqRecordName)) {
-            addErrorMessage(
-                    "Node ID field '%s' in %s '%s' references a table with an inverse key which is not supported.",
-                    field.getName(),
-                    inputKind,
-                    field.getContainerTypeName()
-            );
-            return;
-        }
-
-        var nodeType = schema.getNodeTypeForNodeIdFieldOrThrow(field);
-        var nodeConfig = schema.getNodeConfigurationForNodeIdFieldOrThrow(field);
-        var firstForeignKeyReferencesTargetTable = getTableJavaFieldNameByTableName(foreignKey.getInverseKey().getTable().getName())
-                .orElseThrow()
-                .equalsIgnoreCase(nodeConfig.targetTable().getName());
-
-        if (field.getFieldReferences().size() > 1 || (field.hasFieldReferences() && !firstForeignKeyReferencesTargetTable)) {
-            addErrorMessage(
-                    "Node ID field '%s' in %s '%s' has a reference via table(s) which is not supported on %ss.",
-                    field.getName(),
-                    inputKind,
-                    field.getContainerTypeName(),
-                    inputKind
-            );
-            return;
-        }
-
-        var targetKey =  getPrimaryOrUniqueKeyMatchingFields(nodeConfig.targetTable().getName(), nodeConfig.keyColumnsJavaNames());
-
-        if (targetKey.isPresent() && !targetKey.get().equals(foreignKey.getKey())) {
-            addErrorMessage(
-                    "Node ID field '%s' in %s '%s' uses foreign key '%s' which does not reference the same primary/unique key used for type '%s's node ID. This is not supported.",
-                    field.getName(),
-                    inputKind,
-                    field.getContainerTypeName(),
-                    foreignKey.getName(),
-                    nodeType.getName()
-            );
-        }
-    }
-
-    private boolean isUsedInUpdateMutation(RecordObjectSpecification<?> jooqRecordInput) {
-        var mutation = schema.getMutationType();
-        if (mutation == null) {
-            return false;
-        }
-
-        var usages = mutation.getFields().stream()
-                .filter(ObjectField::hasMutationType)
-                .flatMap(objectField -> {
-                    var inputs = new InputParser(objectField, schema).getJOOQRecords().values();
-                    return inputs.stream()
-                            .map(schema::getInputType)
-                            .filter(it -> it != null && it.getName().equals(jooqRecordInput.getName()))
-                            .map(it -> objectField.getMutationType());
-                }).collect(Collectors.toSet());
-
-        if (usages.isEmpty()) {
-            return false;
-        }
-        return usages.stream().anyMatch(mt -> mt == MutationType.UPDATE || mt == MutationType.UPSERT);
-    }
-
-
-    private void validateNodeIdFieldsInJavaRecordInputs() {
-        schema.getInputTypes().values().stream()
-                .filter(RecordObjectDefinition::hasJavaRecordReference)
-                .flatMap(it -> it.getFields().stream())
-                .filter(GenerationSourceField::hasNodeID)
-                .filter(it -> schema.getNodeTypeForNodeIdField(it).isPresent())
-                .forEach(it -> getJooqRecordClassForNodeIdInputField(it, schema)
-                        .filter(recordClass -> !schema.getNodeTypeForNodeIdFieldOrThrow(it).getTable().getRecordClass().equals(recordClass))
-                        .ifPresent(recordClass -> validateNodeIdReferenceInRecord(
-                                TableReflection.getTableJavaFieldNameForRecordClass(recordClass).orElseThrow(),
-                                it, false, JOOQMapping.fromTable(getTableName(recordClass)))));
-    }
-
-    private void validateTablesAndKeys() {
-        var recordTypes = schema.getRecordTypes();
-        recordTypes
-                .values()
-                .stream()
-                .filter(RecordObjectSpecification::hasTable)
-                .map(RecordObjectSpecification::getTable)
-                .map(JOOQMapping::getMappingName)
-                .filter(it -> !tableExists(it))
-                .forEach(it -> addErrorMessage("No table with name \"%s\" found.", it));
-
-        var allReferences = recordTypes
-                .values()
-                .stream()
-                .flatMap(it -> it.getFields().stream())
-                .flatMap(it -> it.isInput() ? Stream.of(it) : Stream.concat(Stream.of(it), ((ObjectField) it).getArguments().stream()))
-                .flatMap(it -> it.getFieldReferences().stream())
-                .toList();
-
-        allReferences
-                .stream()
-                .filter(FieldReference::hasTable)
-                .map(FieldReference::getTable)
-                .map(JOOQMapping::getMappingName)
-                .filter(it -> !tableExists(it))
-                .forEach(it -> addErrorMessage("No table with name \"%s\" found.", it));
-
-        allReferences
-                .stream()
-                .filter(FieldReference::hasKey)
-                .map(FieldReference::getKey)
-                .map(JOOQMapping::getMappingName)
-                .filter(it -> !keyExists(it))
-                .forEach(it -> addErrorMessage("No key with name \"%s\" found.", it));
-
-        recordTypes
-                .values()
-                .stream()
-                .filter(schema::hasTableObjectForObject)
-                .filter(it -> !schema.isMultiTableInterface(it.getName())) // Don't validate fields on the interface definition since the interface itself has no table
-                .filter(it -> !it.hasJavaRecordReference()) // Don't validate fields on @record types since their data comes from Java records
-                .flatMap(it -> it.getFields().stream())
-                .flatMap(it -> schema.isMultiTableField(it)
-                        ? schema.getTypesFromInterfaceOrUnion(it.getTypeName()).orElse(List.of()).stream().map(o -> new VirtualSourceField(o.getName(), (ObjectField) it))
-                        : Stream.of(it))
-                .filter(field -> !field.hasServiceReference())
-                .filter(field -> schema.isRecordType(field.getTypeName()) || field.hasFieldReferences() || field.createsDataFetcher())
-                .forEach(this::validateReferencePath);
-    }
-
-    private void validateReferencePath(GenerationField field) {
-        var targetTable = getTargetTableForField(field);
-
-        if (targetTable == null) return; // Not a reference
-
-        var sourceTable = schema.getPreviousTableObjectForObject(schema.getRecordTypes().get(field.getContainerTypeName())).getTable().getName();
-        validateReferencePath(field, sourceTable, targetTable);
-    }
-
-    private void validateReferencePath(GenerationField field, String sourceTable, String targetTable) {
-        if (sourceTable.equals(targetTable) && !field.createsDataFetcher() && !field.hasFieldReferences() && !field.hasNodeID()) {
-            return;
-        }
-        for (FieldReference fieldReference : field.getFieldReferences()) {
-            if (fieldReference.hasTableCondition() && !fieldReference.hasKey()) {
-                return;
-            } else if (fieldReference.hasKey()) {
-                String nextTable;
-                String keyName = fieldReference.getKey().getName();
-                var keyTarget = getKeyTargetTableJavaName(keyName).orElse("");
-                if (!keyTarget.equals(sourceTable)) {
-                    nextTable = keyTarget;
-                } else {
-                    nextTable = getKeySourceTableJavaName(keyName).orElse("");
-                }
-                if (nextTable.equals(targetTable)) {
-                    return;
-                } else {
-                    sourceTable = nextTable;
-                }
-            } else if (fieldReference.hasTable()) {
-                String nextTable = fieldReference.getTable().getName();
-                addErrorMessageAndThrowIfNoImplicitPath(field, sourceTable, nextTable);
-                if (nextTable.equals(targetTable)) {
-                    return;
-                } else {
-                    sourceTable = nextTable;
-                }
-            }
-        }
-
-        // Because scalar fields have the whole reference path in the reference directive, validation has completed at this point
-        if (schema.isScalar(field)) {
-            return;
-        }
-
-        addErrorMessageAndThrowIfNoImplicitPath(field, sourceTable, targetTable);
-    }
-
     /**
-     * Walk a reference path from a source table and return the final resolved table name.
-     * Returns empty if the path contains a condition-only element (no table metadata available).
+     * Validate field types separately, as this is called conditionally from {@link ProcessedSchema#validate}.
      */
-    private static Optional<String> resolveReferenceTargetTable(List<FieldReference> references, String sourceTable) {
-        var currentTable = sourceTable;
-        for (FieldReference ref : references) {
-            if (ref.hasKey()) {
-                var resolved = resolveKeyOtherTable(ref.getKey().getName(), currentTable);
-                if (resolved.isEmpty()) return Optional.empty();
-                currentTable = resolved.get();
-            } else if (ref.hasTable()) {
-                currentTable = ref.getTable().getMappingName();
-            } else {
-                return Optional.empty();
-            }
-        }
-        return Optional.of(currentTable);
-    }
-
-    private void addErrorMessageAndThrowIfNoImplicitPath(GenerationField field, String leftTable, String rightTable) {
-        var possibleKeys = getNumberOfForeignKeysBetweenTables(leftTable, rightTable);
-        if (possibleKeys == 1) return;
-
-        var isMultitableField = field instanceof VirtualSourceField;
-
-        addErrorMessageAndThrow("Error on field \"%s\" in type \"%s\": " +
-                        "%s found between tables \"%s\" and \"%s\"%s. Please specify path with the @%s directive.",
-                isMultitableField ? ((VirtualSourceField) field).getOriginalFieldName() : field.getName(),
-                field.getContainerTypeName(),
-                possibleKeys == 0 ? "No foreign key" : "Multiple foreign keys",
-                leftTable,
-                rightTable,
-                isMultitableField ? String.format(" in reference path for type '%s'", field.getTypeName()) : "",
-                isMultitableField ? MULTITABLE_REFERENCE.getName() : REFERENCE.getName()
-        );
-    }
-
-    /**
-     * Returns target table for field determined by target type's table, nodeId directive or 'table' in last item provided in reference directive.
-     * Does not find target table from key only.
-     * @return Name of the target table
-     */
-    private String getTargetTableForField(GenerationField field) {
-        if (schema.isNodeIdField(field)) {
-            return schema.getNodeTypeForNodeIdFieldOrThrow(field).getTable().getName();
-        }
-        if (schema.hasTableObject(field)) {
-            return schema.getObjectOrConnectionNode(field).getTable().getName();
-        }
-
-        if (schema.isScalar(field)) {
-            if (!field.hasFieldReferences()) {
-                return null;
-            }
-            var references = field.getFieldReferences();
-
-            return Optional.ofNullable(references.get(references.size() - 1))
-                    .filter(FieldReference::hasTable)
-                    .map(it -> it.getTable().getName()).orElse(null);
-        }
-        return null;
-    }
-
-    private void validateFieldExists(String tableJavaName, GenerationField field) {
-        if (!tableExists(tableJavaName)) {
-            return;
-        }
-        if (field.isID() && !GeneratorConfig.shouldMakeNodeStrategy()) { // Skip validation for old ID logic
-            return;
-        }
-        if (getJavaFieldName(tableJavaName, field.getUpperCaseName()).isEmpty()) {
-            if (shouldMakeNodeStrategy() && field.isID()) {
-                addErrorMessage("No field with name '%s' found in table '%s' which may be required by %s. " +
-                                "Add %s directive if %s is supposed to be a node ID field.",
-                        field.getUpperCaseName(), tableJavaName, field.formatPath(), NODE_ID.getName(), field.formatPath());
-            } else {
-                addErrorMessage("No field with name '%s' found in table '%s' which is required by %s.",
-                        field.getUpperCaseName(), tableJavaName, field.formatPath());
-            }
-        }
-    }
-
-    private void validateRequiredTableFields() {
-        // Validate objects with table
-        schema.getObjects().values().stream()
-                .filter(RecordObjectDefinition::hasTable)
-                .toList()
-                .forEach(object -> validateTableFieldsExist(object.getTable(), object.getFields()));
-
-        // Validate input objects with table (jOOQ record input)
-        var alreadyValidatedJooqRecordInput = schema.getInputTypes().values().stream()
-                .filter(RecordObjectDefinition::hasTable)
-                .filter(it -> !it.hasJavaRecordReference())
-                .map(input -> {
-                    validateTableFieldsExist(input.getTable(), input.getFields(), false);
-                    return input.getName();
-                })
-                .toList();
-
-        // Validate input on fields returning table object (except already validated jOOQ record input types)
-        allFields.stream()
-                .filter(schema::isObject)
-                .filter(GenerationSourceField::isGenerated)
-                .filter(ObjectField::hasNonReservedInputFields)
-                .filter(it -> schema.getObjectOrConnectionNode(it) != null)
-                .filter(it -> schema.getObjectOrConnectionNode(it).hasTable())
-                .forEach(field -> {
-                    var table = schema.getObjectOrConnectionNode(field).getTable();
-                    var argumentsToValidate = field.getNonReservedArguments().stream()
-                            .filter(it -> !alreadyValidatedJooqRecordInput.contains(it.getTypeName()))
-                            .toList();
-                    validateTableFieldsExist(table, argumentsToValidate, field.hasOverridingCondition());
-                });
-
-        // Validate fields with implicit table from input jOOQ record
-        allFields.stream()
-                .filter(GenerationSourceField::isGenerated)
-                .filter(it -> !it.hasServiceReference())
-                .filter(it -> schema.findInputTables(it).size() == 1)
-                .filter(it -> schema.getObjectOrConnectionNode(it) == null || !schema.getObjectOrConnectionNode(it).hasTable())
-                .forEach(field -> {
-                    var implicitTableFromInput = schema.findInputTables(field).get(0);
-                    List<ObjectField> fields;
-
-                    if (field.hasMutationType()) {
-                        var dataTargetOptional = schema.inferDataTargetForMutation(field);
-                        if (dataTargetOptional.isEmpty()) {
-                            return; // Skip further validation because this is an invalid mutation
-                        }
-
-                        var dataTarget = dataTargetOptional.get();
-                        if (schema.isObject(dataTarget) && schema.getObject(dataTarget).hasTable()) {
-                            return; // Table types are validated above
-                        }
-                        fields = schema.isScalar(dataTarget) ? List.of(dataTarget) : schema.getObject(dataTarget).getFields();
-                    } else if (schema.isObject(field)) {
-                        fields = schema.getObject(field).getFields();
-                    } else if (schema.isScalar(field)) {
-                        fields = List.of(field);
-                    } else {
-                        return;
-                    }
-                    validateTableFieldsExist(implicitTableFromInput, fields);
-                });
-    }
-
-    private void validateTableFieldsExist(JOOQMapping currentTable, Collection<ObjectField> fields) {
-        for (var field : getFieldsInTableContext(fields, currentTable)) {
-            if (schema.isObject(field)) { // Field wrapper object
-                if (field.hasSelectConstruct()) {
-                    continue; // Validated in validateConstructTypeDirective
-                }
-                validateTableFieldsExist(currentTable, schema.getObject(field).getFields());
-                continue;
-            }
-            validateFieldExists(currentTable.getMappingName(), field);
-        }
-    }
-
-    private void validateTableFieldsExist(JOOQMapping currentTable, Collection<? extends InputField> fields, boolean hasOverridingCondition) {
-        for (var field : getFieldsInTableContext(fields, currentTable)) {
-            var hasCondition = hasOverridingCondition || field.hasOverridingCondition();
-            if (schema.isInputType(field)) {
-                var inputType = schema.getInputType(field);
-                if (!inputType.hasTable()) {
-                    validateTableFieldsExist(currentTable, inputType.getFields(), hasCondition);
-                }
-            } else if (!hasCondition) {
-                validateFieldExists(currentTable.getMappingName(), field);
-            }
-        }
-        // Validate that @field columns exist in @reference tables
-        if (!hasOverridingCondition) {
-            fields.stream()
-                    .filter(field -> field.isGenerated() && field.hasFieldReferences() && !field.hasNodeID() && !field.hasOverridingCondition())
-                    .forEach(field ->
-                            resolveReferenceTargetTable(field.getFieldReferences(), currentTable.getMappingName())
-                                    .ifPresent(table -> validateFieldExists(table, field)));
-        }
-    }
-
-    private <T extends GenerationField> List<T> getFieldsInTableContext(Collection<T> fieldList, JOOQMapping currentTable) {
-        return fieldList.stream()
-                .filter(GenerationTarget::isGenerated)
-                .filter(it -> !leavesTableContext(it, currentTable))
-                .toList();
-    }
-
-    private boolean leavesTableContext(GenerationField field, JOOQMapping currentTable) {
-        return field.hasFieldReferences()
-                || (schema.isObject(field) && schema.getObject(field).hasTable() && !schema.getObject(field).getTable().equals(currentTable))
-                || schema.isConnectionObject(field)
-                || schema.isMultiTableField(field)
-                || schema.isInterface(field)
-                || schema.isNodeIdField(field)
-                || field.isExternalField()
-                || field.hasServiceReference()
-                || (schema.isInputType(field) && schema.getInputType(field).hasJavaRecordReference());
-    }
-
-    private void validateExternalMappingReferences() {
-        var referenceSet = GeneratorConfig.getExternalReferences();
-        schema
-                .getEnums()
-                .values()
-                .stream()
-                .filter(EnumDefinition::hasJavaEnumMapping)
-                .map(EnumDefinition::getEnumReference)
-                .filter(e -> !referenceSet.contains(e))
-                .forEach(e -> addErrorMessage("No enum with name '%s' found.", e.getSchemaClassReference()));
-
-        allFields.stream()
-                .filter(ObjectField::hasServiceReference)
-                .map(ObjectField::getExternalMethod)
-                .map(ServiceWrapper::getReference)
-                .filter(e -> !referenceSet.contains(e))
-                .forEach(e -> addErrorMessage("No service with name '%s' found.", e.getSchemaClassReference()));
-
-        allFields
-                .stream()
-                .filter(ObjectField::isGenerated)
-                .filter(ObjectField::hasCondition)
-                .map(ObjectField::getCondition)
-                .map(SQLCondition::getReference)
-                .filter(e -> !referenceSet.contains(e))
-                .forEach(e -> addErrorMessage("No condition with name '%s' found.", e.getSchemaClassReference()));
-    }
-
-    private void validateServiceMethods() {
-        allFields.stream()
-                .filter(ObjectField::hasServiceReference)
-                .map(GenerationSourceField::getExternalMethod)
-                .filter(it -> GeneratorConfig.getExternalReferences().contains(it.getReference()))
-                .filter(it -> getAllOverloads(it).isEmpty())
-                .forEach(it -> addErrorMessage("Service reference with name '%s' does not contain a method named '%s'.", GeneratorConfig.getExternalReferences().getClassFrom(it.getReference()), it.getReference().getMethodName()));
-    }
-
-    private void validateServiceInputTypes() {
-        allFields.stream()
-                .filter(ObjectField::hasServiceReference)
-                .forEach(field -> field.getNonReservedArguments()
-                        .stream()
-                        .filter(schema::isInputType)
-                        .filter(arg -> !schema.hasRecord(arg))
-                        .forEach(arg -> addErrorMessage(
-                                "Input type '%s' is used as an argument on service field '%s.%s', but has neither the @%s nor the @%s directive. "
-                                        + "Input types on @%s operations must have one of these directives.",
-                                arg.getTypeName(),
-                                field.getContainerTypeName(),
-                                field.getName(),
-                                TABLE.getName(),
-                                RECORD.getName(),
-                                SERVICE.getName()
-                        ))
-                );
-    }
-
-    private void validateServiceMethodParameterTypes() {
-        allFields.stream()
-                .filter(ObjectField::hasServiceReference)
-                .filter(field -> field.getNonReservedArguments()
-                        .stream()
-                        .filter(schema::isInputType)
-                        .allMatch(schema::hasRecord))
-                .forEach(this::checkServiceParameterTypes);
-    }
-
-    private void checkServiceParameterTypes(ObjectField field) {
-        var serviceWrapper = field.getExternalMethod();
-        var overloads = getAllOverloads(serviceWrapper);
-        if (overloads.isEmpty()) {
-            return;
-        }
-
-        var parser = new InputParser(field, schema);
-        var resolverKeyOffset = field.isRootField() ? 0 : 1;
-        var expectedParamCount = resolverKeyOffset + parser.getMethodInputNames(true, true, true).size();
-
-        var candidates = overloads.stream()
-                .filter(m -> m.getParameterTypes().length == expectedParamCount)
-                .toList();
-
-        if (candidates.isEmpty()) {
-            addErrorMessage(
-                    "Service field '%s.%s' maps to %d method parameter(s) but there is no overload of '%s' with that parameter count. Available overloads:\n%s",
-                    field.getContainerTypeName(),
-                    field.getName(),
-                    expectedParamCount,
-                    serviceWrapper.getMethodName(),
-                    formatOverloads(overloads, serviceWrapper.getMethodName())
-            );
-            return;
-        }
-
-        var allMethodInputs = parser.getMethodInputNames(false, false, false);
-
-        for (var entry : parser.getRecords().entrySet()) {
-            var varName = entry.getKey();
-            var inputField = entry.getValue();
-
-            var posInInputs = allMethodInputs.indexOf(VariablePrefix.inputPrefix(varName));
-            if (posInInputs < 0) {
-                continue;
-            }
-            var methodParamIndex = resolverKeyOffset + posInInputs;
-
-            var inputDef = schema.getInputType(inputField);
-            if (inputDef == null || !inputDef.hasRecordReference()) {
-                continue;
-            }
-            var inputClass = inputDef.getRecordReference();
-            if (inputClass == null) {
-                continue;
-            }
-
-            var inputIsListed = inputField.isIterableWrapped();
-            if (candidates.stream().noneMatch(o -> isParameterCompatible(o, methodParamIndex, inputClass, inputIsListed))) {
-                addErrorMessage(
-                        "Argument '%s' on service field '%s.%s' has input type '%s' which maps to '%s', "
-                                + "but there is no overload of '%s' that accepts this. Available overloads:\n%s",
-                        inputField.getName(),
-                        field.getContainerTypeName(),
-                        field.getName(),
-                        inputField.getTypeName(),
-                        formatExpectedListableType(inputClass, inputIsListed),
-                        serviceWrapper.getMethodName(),
-                        formatOverloads(candidates, serviceWrapper.getMethodName())
-                );
-            }
-        }
-    }
-
-    private boolean isParameterCompatible(Method method, int paramIndex, Class<?> inputClass, boolean inputIsListed) {
-        var paramTypes = method.getParameterTypes();
-        if (paramIndex >= paramTypes.length) {
-            return false;
-        }
-        if (inputIsListed != List.class.isAssignableFrom(paramTypes[paramIndex])) {
-            return false;
-        }
-        var effectiveType = getEffectiveParameterType(method, paramIndex, inputIsListed);
-        return effectiveType != null && effectiveType.isAssignableFrom(inputClass);
-    }
-
-    private List<Method> getAllOverloads(CodeReferenceWrapper serviceWrapper) {
-        return GeneratorConfig.getExternalReferences().getMethodsFrom(serviceWrapper.getReference());
-    }
-
-    private Class<?> getEffectiveParameterType(Method method, int paramIndex, boolean isListed) {
-        var paramTypes = method.getParameterTypes();
-        if (paramIndex >= paramTypes.length) {
-            return null;
-        }
-        if (!isListed) {
-            return paramTypes[paramIndex];
-        }
-        var genericTypes = method.getGenericParameterTypes();
-        if (paramIndex < genericTypes.length
-                && genericTypes[paramIndex] instanceof ParameterizedType pt
-                && pt.getActualTypeArguments().length > 0
-                && pt.getActualTypeArguments()[0] instanceof Class<?> elementType) {
-            return elementType;
-        }
-        return null;
-    }
-
-    private void validateMultitableFieldsOutsideRoot() {
-        allFields.stream()
-                .filter(GenerationSourceField::isGenerated)
-                .filter(it -> !it.isRootField())
-                .filter(it -> schema.isObjectWithPreviousTableObject(it.getContainerTypeName()))
-                .filter(schema::isMultiTableField)
-                .forEach(field -> {
-                    if (!field.createsDataFetcher()) {
-                        addErrorMessage("%s is a multitable field outside root, but is missing the %s directive. " +
-                                        "Multitable queries outside root is only supported for resolver fields.",
-                                field.formatPath(), SPLIT_QUERY.getName()
-                        );
-                    }
-
-                    if (field.hasFieldReferences()) {
-                        addErrorMessage("%s has the %s directive which is not supported on multitable queries. Use %s directive instead.",
-                                field.formatPath(), REFERENCE.getName(), MULTITABLE_REFERENCE.getName()
-                        );
-                    }
-                });
-    }
-
-    private void validateMultitableTypeTables() {
-        schema
-                .getObjects()
-                .values()
-                .stream()
-                .filter(Objects::nonNull)
-                .flatMap(it -> it.getFields().stream())
-                .filter(schema::isMultiTableField)
-                .filter(it -> it.createsDataFetcher() || it.isRootField())
-                .filter(it -> !it.hasServiceReference())
-                .filter(it -> !it.getName().equals(FEDERATION_ENTITIES_FIELD.getName()))
-                .filter(it -> !it.getTypeName().equals(NODE_TYPE.getName()))
-                .filter(it -> !it.getTypeName().equals(ERROR_TYPE.getName()))
-                .forEach((field) -> {
-                    var multitableName = field.getTypeName();
-                    var typesMissingTable = schema
-                            .getTypesFromInterfaceOrUnion(multitableName).orElse(List.of())
-                            .stream()
-                            .filter(it -> !it.hasTable())
-                            .map(AbstractObjectDefinition::getName)
-                            .collect(Collectors.joining("', '"));
-                    if (!typesMissingTable.isEmpty()) {
-                        addErrorMessage(MISSING_TABLE_ON_MULTITABLE, typesMissingTable, field.formatPath(), field.getTypeName());
-                    }
-                });
-    }
-
-    private void validateSingleTableInterfaceDefinitions() {
-        schema.getInterfaces().forEach((name, interfaceDefinition) -> {
-                    if (name.equalsIgnoreCase(NODE_TYPE.getName()) || name.equalsIgnoreCase(ERROR_TYPE.getName())) return;
-
-                    var implementations = schema.getTypesFromInterfaceOrUnion(interfaceDefinition).orElse(List.of());
-
-                    if (interfaceDefinition.hasDiscriminator() == interfaceDefinition.isMultiTableInterface()) {
-                        addErrorMessage(
-                                String.format("'%s' and '%s' directives on interfaces must be used together. " +
-                                                "Interface '%s' is missing '%s' directive.",
-                                        DISCRIMINATE.getName(), TABLE.getName(), name,
-                                        interfaceDefinition.hasTable() ? DISCRIMINATE.getName() : TABLE.getName()));
-                    }
-
-                    if (!interfaceDefinition.isMultiTableInterface()) { // Single table interface with discriminator
-                        Optional<?> discriminatorField = getField(
-                                interfaceDefinition.getTable().getName(),
-                                interfaceDefinition.getDiscriminatorFieldName()
-                        );
-                        if (discriminatorField.isEmpty()) {
-                            addErrorMessage(
-                                    String.format("Interface '%s' has discriminating field set as '%s', but the field " +
-                                                    "does not exist in table '%s'.",
-                                            name, interfaceDefinition.getDiscriminatorFieldName(), interfaceDefinition.getTable().getName()));
-                        } else {
-                            Optional<Class<?>> fieldType = getFieldType(
-                                    interfaceDefinition.getTable().getName(),
-                                    interfaceDefinition.getDiscriminatorFieldName()
-                            );
-                            if (fieldType.isEmpty() || !fieldType.get().equals(String.class)) {
-                                addErrorMessage(
-                                        String.format("Interface '%s' has discriminating field set as '%s', but the field " +
-                                                        "does not return a string type, which is not supported.",
-                                                name, interfaceDefinition.getDiscriminatorFieldName()));
-                            }
-                        }
-
-                        implementations.forEach(impl -> {
-                            if (!impl.hasDiscriminator()) {
-                                addErrorMessage(
-                                        String.format("Type '%s' is missing '%s' directive in order to implement interface '%s'.",
-                                                impl.getName(), DISCRIMINATOR.getName(), name));
-                            }
-                            if (impl.hasTable() && interfaceDefinition.hasTable() && !impl.getTable().equals(interfaceDefinition.getTable())) {
-                                addErrorMessage(
-                                        String.format("Interface '%s' requires implementing types to have table '%s', " +
-                                                        "but type '%s' has table '%s'.",
-                                                name, interfaceDefinition.getTable().getName(), impl.getName(), impl.getTable().getName()));
-                            }
-
-                            impl.getFields()
-                                    .stream()
-                                    .filter(it -> interfaceDefinition.hasField(it.getName()))
-                                    .forEach(it -> {
-                                        var fieldInInterface = interfaceDefinition.getFieldByName(it.getName());
-                                        String sharedErrorMessage = "Overriding '%s' configuration in types implementing " +
-                                                "a single table interface is not currently supported, and must be identical " +
-                                                "with interface. Type '%s' has a configuration mismatch on field '%s' from the interface '%s'.";
-
-                                        if (it.hasCondition() != fieldInInterface.hasCondition()
-                                                || (it.hasCondition() && !it.getCondition().equals(fieldInInterface.getCondition()))) {
-                                            addErrorMessage(
-                                                    sharedErrorMessage,
-                                                    CONDITION.getName(), impl.getName(), it.getName(), name
-                                            );
-                                        }
-
-                                        if (!it.getFieldReferences().equals(fieldInInterface.getFieldReferences())) {
-                                            addErrorMessage(
-                                                    sharedErrorMessage,
-                                                    REFERENCE.getName(), impl.getName(), it.getName(), name
-                                            );
-                                        }
-                                    });
-                        });
-
-                        // Check for conflicts in "shared" fields which are not in the interface definition
-                        implementations.stream()
-                                .map(AbstractObjectDefinition::getFields)
-                                .flatMap(Collection::stream)
-                                .filter(it -> !interfaceDefinition.hasField(it.getName()))
-                                .collect(groupingBy(ObjectField::getName))
-                                .entrySet()
-                                .stream()
-                                .filter(it -> it.getValue().size() > 1)
-                                .forEach(entry -> {
-                                    var fields = entry.getValue();
-                                    var first = fields.get(0);
-
-                                    fields.stream().skip(1).forEach(field -> {
-                                        String sharedErrorMessage = "Different configuration on fields in types implementing the same single table interface is currently not supported. " +
-                                                "Field '%s' occurs in two or more types implementing interface '%s', but there is a mismatch between the configuration of the '%s' directive.";
-                                        if (!field.getUpperCaseName().equals(first.getUpperCaseName())) {
-                                            addErrorMessage(
-                                                    sharedErrorMessage,
-                                                    entry.getKey(), interfaceDefinition.getName(), FIELD.getName()
-                                            );
-                                        }
-
-                                        if (!field.getFieldReferences().equals(first.getFieldReferences())) {
-                                            addErrorMessage(
-                                                    sharedErrorMessage,
-                                                    entry.getKey(), interfaceDefinition.getName(), REFERENCE.getName());
-                                        }
-
-                                        if (!(field.hasCondition() == first.hasCondition()
-                                                && (!field.hasCondition() || field.getCondition().equals(first.getCondition())))) {
-                                            addErrorMessage(
-                                                    sharedErrorMessage,
-                                                    entry.getKey(), interfaceDefinition.getName(), CONDITION.getName());
-                                        }
-                                    });
-                                });
-                    }
-                }
-        );
-
-        schema.getObjects().values().stream()
-                .filter(objectDefinition -> !objectDefinition.getImplementedInterfaces().isEmpty() || objectDefinition.hasDiscriminator())
-                .forEach(objectDefinition -> {
-                    var implementedInterfaces = objectDefinition.getImplementedInterfaces();
-
-                    var singleTableInterfaces = implementedInterfaces.stream()
-                            .filter(it -> !it.equals(NODE_TYPE.getName()))
-                            .filter(schema::isSingleTableInterface)
-                            .toList();
-
-                    if (singleTableInterfaces.isEmpty() && objectDefinition.hasDiscriminator()) {
-                        addErrorMessage(
-                                String.format("Type '%s' has discriminator, but doesn't implement any interfaces requiring it.", objectDefinition.getName())
-                        );
-                    }
-                });
-    }
-
-    private void validateInterfacesReturnedInFields() {
-        allFields.stream()
-                .filter(ObjectField::isGeneratedWithResolver)
-                .filter(schema::isInterface)
-                .forEach(field -> {
-                            var typeName = field.getTypeName();
-                            var name = Optional
-                                    .ofNullable(schema.getObjectOrConnectionNode(typeName))
-                                    .map(AbstractObjectDefinition::getName)
-                                    .orElse(typeName);
-                            var isSingleTable = schema.isSingleTableInterface(field);
-                            if (!field.isRootField() && (isSingleTable || name.equals(NODE_TYPE.getName()))) {
-                                addErrorMessage("interface (%s) returned in non root object. This is not fully " +
-                                        "supported. Use with care", name);
-                            }
-
-                            if (name.equalsIgnoreCase(NODE_TYPE.getName())) {
-                                isTrue(
-                                        field.getArguments().size() == 1,
-                                        "Only exactly one input field is currently supported for fields returning the '%s' interface. " +
-                                                "%s has %s input fields", NODE_TYPE.getName(), field.formatPath(), field.getArguments().size()
-                                );
-                                isTrue(
-                                        !field.isIterableWrapped(),
-                                        "Generating fields returning a list of '%s' is not supported. " +
-                                                "'%s' must return only one %s", name, field.getName(), field.getTypeName()
-                                );
-                            } else {
-                                schema.getObjects()
-                                        .values()
-                                        .stream()
-                                        .filter(it -> it.implementsInterface(schema.getInterface(name).getName()))
-                                        .forEach(implementation -> {
-                                            if (!implementation.hasTable() && isSingleTable) {
-                                                addErrorMessage("Interface '%s' is returned in field '%s', but type '%s' " +
-                                                        "implementing '%s' does not have table set. This is not supported.", name, field.getName(), implementation.getName(), name);
-                                            } else if (implementation.hasTable() && !tableHasPrimaryKey(implementation.getTable().getName())) {
-                                                addErrorMessage("Interface '%s' is returned in field '%s', but implementing type '%s' " +
-                                                        "has table '%s' which does not have a primary key. This is not supported.", name, field.getName(), implementation.getName(), implementation.getTable().getName());
-                                            }
-                                        });
-                            }
-                        }
-                );
-    }
-
-    private void validateTypesUsingNodeInterfaceWithoutNodeDirective() {
-        if (!schema.nodeExists() ||
-                schema.getQueryType() == null ||
-                schema.getQueryType().getFieldByName(uncapitalize(NODE_TYPE.getName())) == null ||
-                schema.getQueryType().getFieldByName(uncapitalize(NODE_TYPE.getName())).isExplicitlyNotGenerated()) {
-            return;
-        }
-
-        var records = schema
-                .getObjects()
-                .values()
-                .stream()
-                .filter(it -> it.implementsInterface(NODE_TYPE.getName()) && it.hasTable() && !it.hasNodeDirective())
-                .collect(groupingBy(
-                        it -> it.getTable().getName(), Collectors.mapping(ObjectDefinition::getName, Collectors.toSet())));
-
-        records.forEach((tableName, schemaTypes) -> {
-            if (schemaTypes.size() > 1) {
-                addErrorMessage(
-                        "Multiple types (%s) implement the %s interface and refer to the same table %s. This is not supported.",
-                        String.join(", ", schemaTypes), NODE_TYPE.getName(), tableName);
-            }
-        });
-    }
-
-    // TODO: It seems we are now handling/validating lists that are directly below parent lists. But what about
-    //  validation of nested lists where a list is not directly below parent list, but are deeper down the hierarchy?
-    private void validateInputFields() {
-        var oneLayerFlattenedFields = allFields
-                .stream()
-                .filter(schema::hasTableObject)
-                .flatMap(it -> it.getNonReservedArguments().stream())
-                .filter(it -> !schema.hasRecord(it) && !it.hasOverridingCondition())
-                .filter(AbstractField::isIterableWrapped)
-                .filter(schema::isInputType)
-                .toList();
-
-        for (var field : oneLayerFlattenedFields) {
-            var messageStart = String.format("Argument '%s' is a collection of InputFields ('%s') type.", field.getName(), field.getTypeName());
-            var inputDefinitionFields = schema.getInputType(field).getFields();
-
-            inputDefinitionFields.stream().filter(AbstractField::isIterableWrapped).findFirst().ifPresent(it -> addErrorMessageAndThrow(
-                    "%s Fields returning collections: '%s' are not supported on such types (used for generating condition tuples)",
-                    messageStart,
-                    it.getName()
-            ));
-
-            var optionalFields = inputDefinitionFields
-                    .stream()
-                    .filter(AbstractField::isNullable)
-                    .map(AbstractField::getName)
-                    .collect(Collectors.toList());
-            if (!optionalFields.isEmpty()) {
-                addErrorMessage(
-                        String.format(
-                                "%s Optional fields on such types are not supported. The following fields will be " +
-                                        "treated as mandatory in the resulting, generated condition tuple: '%s'",
-                                messageStart,
-                                String.join("', '", optionalFields)
-                        )
-                );
-            }
-        }
-    }
-
     public void validateObjectFieldTypes() {
-        schema.getRecordTypes()
-                .values()
-                .stream()
-                .map(this::validateFieldTypes)
-                .filter(it -> !it.isEmpty())
-                .forEach(ValidationHandler::addErrorMessage);
-
-        throwIfErrors();
-
-    }
-
-    private String validateFieldTypes(RecordObjectSpecification<? extends GenerationField> object) {
-        var objectName = object.getName();
-        return object
-                .getFields()
-                .stream()
-                .flatMap(it -> validateFieldType(objectName, it).stream())
-                .filter(it -> !it.isEmpty())
-                .collect(Collectors.joining("\n"));
-    }
-
-    private List<String> validateFieldType(String objectName, GenerationField field) {
-        return checkTypeExists(field.getTypeName(), Set.of()) // May contain more than one if type is union.
-                .entrySet()
-                .stream()
-                .map(it -> String.format("Field \"%s\" within schema type \"%s\" has invalid type \"%s\" (or an union containing it). Closest type matches found by levenshtein distance are:\n%s", field.getName(), objectName, it.getKey(), it.getValue()))
-                .collect(Collectors.toList());
-    }
-
-    private Map<String, String> checkTypeExists(String typeName, Set<String> seenTypes) {
-        if (seenTypes.contains(typeName) || schema.isScalar(typeName)) { // Check seen types in case there are some strange Union recursions.
-            return Map.of();
-        }
-
-        if (schema.isUnion(typeName)) {
-            return schema
-                    .getUnion(typeName)
-                    .getFieldTypeNames()
-                    .stream()
-                    .flatMap(it -> checkTypeExists(it, Stream.concat(seenTypes.stream(), Stream.of(typeName)).collect(Collectors.toSet())).entrySet().stream())
-                    .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-        }
-
-        if (schema.isType(typeName)) {
-            return Map.of();
-        }
-
-        // If we make it here, no valid type could be identified.
-        var similarTypes = findSimilarStringsWithDistance(typeName, schema.getAllValidFieldTypeNames().stream(), 12);
-        var formatted = similarTypes.entrySet().stream()
-                .map(e -> e.getKey() + " - " + e.getValue())
-                .collect(Collectors.joining(", "));
-        return Map.of(typeName, formatted);
-    }
-
-    private void checkPaginationSpecs(List<ObjectField> fields) {
-        for (var field : fields) {
-            var hasConnectionSuffix = field.getTypeName().endsWith(SCHEMA_CONNECTION_SUFFIX.getName());
-            if (hasConnectionSuffix && !field.hasRequiredPaginationFields()) {
-                addErrorMessage(
-                        "Type %s ending with the reserved suffix 'Connection' must have either " +
-                                "forward(first and after fields) or backwards(last and before fields) pagination, " +
-                                "yet neither was found.", field.getTypeName()
-                );
-            }
-        }
-    }
-
-    private void validateMutationDirectives() {
-        var mutation = schema.getMutationType();
-        if (mutation == null) {
-            return;
-        }
-
-        var mutations = mutation
-                .getFields()
-                .stream()
-                .filter(ObjectField::isGeneratedWithResolver)
-                .toList();
-        if (mutations.isEmpty()) {
-            return;
-        }
-
-        mutations
-                .stream()
-                .filter(it -> !it.hasMutationType() && !it.hasServiceReference())
-                .forEach(it -> addErrorMessage("Mutation '%s' is set to generate, but has neither a service nor mutation type set.", it.getName()));
-        mutations
-                .stream()
-                .filter(it -> it.hasMutationType() && !new InputParser(it, schema).hasJOOQRecords())
-                .forEach(it -> addErrorMessage("Mutations must have at least one table attached when generating resolvers with queries. Mutation '%s' has no tables attached.", it.getName()));
-    }
-
-    private void validateMutationRequiredFields() {
-        var mutation = schema.getMutationType();
-        if (mutation == null) {
-            return;
-        }
-
-        mutation
-                .getFields()
-                .stream()
-                .filter(ObjectField::isGeneratedWithResolver)
-                .filter(ObjectField::hasMutationType)
-                .forEach(target -> {
-                    validateRecordRequiredFields(target);
-                    validateMutationWithReturning(target);
-                    new InputParser(target, schema).getJOOQRecords().values().forEach(inputField -> checkMutationIOFields(inputField, target));
-                });
-    }
-
-    private void validateRecursiveRecordInputs() {
-        var mutation = schema.getMutationType();
-        var mutations = mutation != null ? mutation.getFields().stream() : Stream.<ObjectField>of();
-        var query = schema.getQueryType();
-        var queries = query != null ? query.getFields().stream() : Stream.<ObjectField>of();
-
-        Stream
-                .concat(queries, mutations)
-                .filter(ObjectField::hasServiceReference)
-                .flatMap(it -> it.getArguments().stream())
-                .filter(schema::isInputType)
-                .forEach(it -> validateRecursiveRecordInputs(it, false, 0));
-    }
-
-    private void validateOnlyOneInputRecordInputWhenNoTypeTableIsPresent() {
-        var mutation = schema.getMutationType();
-        var mutations = mutation != null ? mutation.getFields().stream() : Stream.<ObjectField>of();
-        var query = schema.getQueryType();
-        var queries = query != null ? query.getFields().stream() : Stream.<ObjectField>of();
-
-        Stream
-                .concat(queries, mutations)
-                .filter(ObjectField::isGeneratedWithResolver)
-                .filter(it -> !it.hasServiceReference())
-                .filter(it -> !it.getTypeName().equals(FEDERATION_SERVICE_TYPE.getName()))
-                .filter(it -> !it.getTypeName().equals(FEDERATION_ENTITY_UNION.getName()))
-                .filter(it -> !schema.isInterface(it))
-                .filter(it -> !schema.isUnion(it))
-                .filter(it -> schema.isScalar(it) || schema.isRecordType(it))
-                .filter(it -> !schema.nextTypeTableExists(it, new HashSet<>()))
-                .collect(Collectors.toMap(it -> it.getContainerTypeName() + "." + it.getName(), it -> schema.findInputTables(it).size()))
-                .entrySet()
-                .stream()
-                .filter(it -> it.getValue() != 1)
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue))
-                .forEach((key, value) -> addErrorMessage(
-                        "%s is a field of a type without a table, and has %s potential input records to use as a source for the table in queries. In such cases, there must be exactly one input table so that it can be resolved unambiguously.",
-                        key,
-                        value
-                ));
-    }
-
-    private void validateRecursiveRecordInputs(InputField field, boolean wasRecord, int recursion) {
-        recursionCheck(recursion);
-
-        var input = schema.getInputType(field);
-        if (input == null) {
-            return;
-        }
-
-        var hasTableOrRecordReference = input.hasTable() || input.hasJavaRecordReference();
-
-        if (field.isIterableWrapped() && wasRecord && !hasTableOrRecordReference) {
-            addErrorMessage(
-                    String.format(
-                            "Field %s with Input type %s is iterable, but has no record mapping set. Iterable Input types within records without record mapping can not be mapped to a single field in the surrounding record.",
-                            field.getName(),
-                            input.getName()
-                    )
-            );
-        }
-
-        input.getFields().forEach(it -> validateRecursiveRecordInputs(it, wasRecord || hasTableOrRecordReference, recursion + 1));
-    }
-
-    //Check input and payload("output") fields
-    private void checkMutationIOFields(InputField inputField, ObjectField objectField) {
-        if(!schema.isObject(objectField))
-            return;
-
-        var objectFieldErrors = new InputParser(objectField, schema).getAllErrors();
-        var payloadContainsIterableField = objectField.isIterableWrapped() ||
-                schema.getObject(objectField)
-                        .getFields()
-                        .stream()
-                        .filter(field -> !objectFieldErrors.contains(field))
-                        .anyMatch(AbstractField::isIterableWrapped);
-
-        if (!inputField.isIterableWrapped() && payloadContainsIterableField) {
-            addWarningMessage(
-                    "Mutation %s with Input %s is not defined as a list while Payload type %s contains " +
-                            "a list",
-                    objectField.getName(), inputField.getTypeName(), objectField.getTypeName());
-        } else if (inputField.isIterableWrapped() && !payloadContainsIterableField) {
-            addWarningMessage(
-                    "Mutation %s with Input %s is defined as a list while Payload type %s does not " +
-                            "contain a list",
-                    objectField.getName(), inputField.getTypeName(), objectField.getTypeName());
-        }
-    }
-
-    private void validateRecordRequiredFields(ObjectField target) {
-        var mutationType = target.getMutationType();
-        if (mutationType.equals(MutationType.INSERT) || mutationType.equals(MutationType.UPSERT)) {
-            var recordInputs = new InputParser(target, schema).getJOOQRecords().values();
-            if (recordInputs.isEmpty()) {
-                addErrorMessage("Mutation "
-                        + target.getName()
-                        + " is set as an " + lowerCase(mutationType.name()) + " operation, but does not link any input to tables.");
-            }
-
-            recordInputs.forEach(this::checkRequiredFields);
-        }
-    }
-
-    private void validateMutationWithReturning(ObjectField field) {
-        boolean isDeleteWithReturning = schema.isDeleteMutationWithReturning(field);
-        boolean isInsertWithReturning = schema.isInsertMutationWithReturning(field);
-
-        if (!isDeleteWithReturning && !isInsertWithReturning) {
-            return;
-        }
-        /* Validate output */
-        var dataField = schema.inferDataTargetForMutation(field);
-
-        if (dataField.isEmpty()) {
-            addErrorMessage("Cannot find correct field to output data to after mutation for field %s.", field.formatPath());
-            return;
-        }
-
-        /* Validate input */
-        var recordInputs = new InputParser(field, schema).getJOOQRecords().values(); // TODO: support non-jOOQ record inputs
-        if (recordInputs.isEmpty()) {
-            addErrorMessage("Field %s is a generated %s mutation, but does not link any input to tables.", field.formatPath(), field.getMutationType());
-            return;
-        } else if (recordInputs.size() != 1) {
-            addErrorMessage("Field %s is a generated %s mutation, but has multiple input records. This is not supported.", field.formatPath(), field.getMutationType());
-            return;
-        }
-
-        var input = recordInputs.stream().findFirst().orElseThrow();
-
-        if (isInsertWithReturning && !shouldMakeNodeStrategy() && schema.getInputType(input).getFields().stream().anyMatch(AbstractField::isID)) {
-            addErrorMessage("%s is a generated %s field with ID input, but this is only supported with node ID strategy enabled.",
-                    field.formatPath(),
-                    field.getMutationType()
-            );
-            return;
-        }
-
-        var inputTable = schema.getRecordType(input).getTable();
-        var outputTable = schema.isScalar(dataField.get()) ? inputTable : schema.getRecordType(dataField.get()).getTable();
-
-        if (!inputTable.equals(outputTable)) {
-            addErrorMessage("Mutation field %s has a mismatch between input and output tables. Input table is '%s', and output table is '%s'.",
-                    field.formatPath(), inputTable.getMappingName(), outputTable.getMappingName());
-            return;
-        }
-
-        if (input.isNullable() || input.isIterableWrappedWithNullableElement()) {
-            addErrorMessage("Field %s is a generated %s mutation, but has nullable input. This is not supported. Consider changing the input type from '%s' to '%s'.",
-                    field.formatPath(), field.getMutationType(),
-                    input.formatGraphQLSchemaType(),
-                    input.isIterableWrapped() ? "[" + input.getTypeName() + "!]!" : input.getTypeName() + "!");
-        }
-
-        if (isDeleteWithReturning) {
-            validateDeleteMutation(field, dataField.orElse(null), input);
-        }
-    }
-
-    private void validateDeleteMutation(ObjectField field, ObjectField dataField, InputField input) {
-        var inputType = schema.getInputType(input);
-        var targetTable = inputType.getTable();
-
-        /* Validate that there are no references in output */
-        if (dataField != null && schema.isRecordType(dataField)){
-            var subqueryReferenceFields = findSubqueryReferenceFieldsForTableObject(dataField, targetTable, 0);
-            if (!subqueryReferenceFields.isEmpty()) {
-                addErrorMessage(
-                        "Mutation field %s has references returned from the data field. " +
-                                "This is not supported for %s mutations. Found reference fields are: %s",
-                        field.formatPath(), MutationType.DELETE, String.join(", ", subqueryReferenceFields.stream().map(GenerationField::formatPath).toList()));
-            }
-        }
-
-        /* Validate that there is non-nullable input matching a PK/UK */
-        var idFields = inputType.getFields().stream()
-                .filter(it -> it.isID() || schema.isNodeIdField(it))
-                .filter(it -> !it.hasFieldReferences())
-                .filter(it -> !schema.isNodeIdField(it) || schema.getNodeTypeForNodeIdField(it).map(n -> n.getTable().equals(targetTable)).orElse(false))
-                .filter(it -> schema.isNodeIdField(it) || it.getUpperCaseName().equalsIgnoreCase(GraphQLReservedName.NODE_ID.getName()))
-                .toList();
-
-        var requiredIdOrNodeIdFields = idFields.stream()
-                .filter(AbstractField::isNonNullable)
-                .toList();
-
-        if (requiredIdOrNodeIdFields.isEmpty()) {
-            var possibleFixes = new ArrayList<>(idFields.stream().map(GenerationSourceField::formatPath).map(it -> "Make " + it + " non-nullable.").toList());
-            if (schema.getNodeTypesWithTable(targetTable).size() == 1) {
-                possibleFixes.add(String.format("Add non-nullable node ID input field for type '%s'.",
-                        schema.getNodeTypesWithTable(targetTable).stream().findFirst().get().getName())
-                );
-            }
-
-            var keys = getPrimaryAndUniqueKeysForTable(targetTable.getName());
-            var nonNullableInputFields = inputType.getFields().stream()
-                    .filter(AbstractField::isNonNullable)
-                    .toList();
-
-            for (var key : keys) {
-                var keyFields = getJavaFieldNamesForKey(targetTable.getName(), key);
-                var keyFieldToInputFieldMatches = keyFields.stream()
-                        .collect(Collectors.toMap(
-                                kf -> kf,
-                                kf -> nonNullableInputFields.stream()
-                                        .filter(f -> f.getUpperCaseName().equalsIgnoreCase(kf))
-                                        .toList()
-                        ));
-
-                if (keyFieldToInputFieldMatches.entrySet().stream().anyMatch(e -> e.getValue().isEmpty())) {
-                    possibleFixes.add(String.format("Add non-nullable input fields for %s to match PK/UK '%s'.",
-                            String.join(", ", keyFieldToInputFieldMatches.entrySet().stream().filter(it -> it.getValue().isEmpty()).map(Map.Entry::getKey).toList()),
-                            key.getName())
-                    );
-                } else {
-                    return;
-                }
-            }
-
-            addErrorMessage("Mutation field %s is a generated delete mutation, but does not have input with non-nullable fields corresponding to a PK/UK of the table. %s%s",
-                    field.formatPath(),
-                    !possibleFixes.isEmpty() ? "\nPossible fix(es):\n* " : "",
-                    String.join("\n* ", possibleFixes)
-            );
-        }
-
-    }
-
-    private Set<GenerationField> findSubqueryReferenceFieldsForTableObject(GenerationField target, JOOQMapping targetTable, int recursion) {
-        recursionCheck(recursion);
-
-        if (target.hasFieldReferences() || isNodeIdReferenceField(target, schema, targetTable)) {
-            return Set.of(target);
-        }
-
-        var result = new HashSet<GenerationField>();
-
-        if (schema.isRecordType(target)) {
-            var recordType = schema.getRecordType(target);
-
-            // Implicit reference
-            if (recordType.hasTable() && !recordType.getTable().equals(targetTable)) {
-                return Set.of(target);
-            }
-
-            recordType.getFields()
-                    .stream()
-                    .filter(GenerationTarget::isGenerated)
-                    .filter(it -> !it.createsDataFetcher())
-                    .forEach(f -> {
-                        if (f.hasFieldReferences()) {
-                            result.add(f);
-                        } else {
-                            result.addAll(findSubqueryReferenceFieldsForTableObject(f, targetTable, recursion + 1));
-                        }
-                    });
-        }
-        return result;
-    }
-
-    protected void checkRequiredFields(InputField recordInput) {
-        var inputObject = schema.getInputType(recordInput);
-        var tableName = inputObject.getTable().getMappingName();
-
-        var requiredDBFields = getRequiredFields(tableName)
-                .stream()
-                .map(String::toUpperCase)
-                .filter(it -> !tableFieldHasDefaultValue(tableName, it)) // No need to complain when it has a default set. Note that this does not work for views.
-                .collect(Collectors.toList());
-        var recordFieldNames =
-                inputObject.getFields()
-                        .stream()
-                        .map(InputField::getUpperCaseName)
-                        .collect(Collectors.toSet());
-        checkRequiredFieldsExist(MISSING_FIELD, recordFieldNames, requiredDBFields, recordInput);
-
-        var requiredRecordFieldNames =
-                inputObject.getFields()
-                        .stream()
-                        .filter(AbstractField::isNonNullable)
-                        .map(InputField::getUpperCaseName)
-                        .collect(Collectors.toSet());
-        checkRequiredFieldsExist(MISSING_NON_NULLABLE, requiredRecordFieldNames, requiredDBFields, recordInput);
-    }
-
-    protected void checkRequiredFieldsExist(InputTableError message, Set<String> actualFields, List<String> requiredFields, InputField recordInput) {
-        if (!actualFields.containsAll(requiredFields)) {
-            var missingFields = requiredFields.stream().filter(it -> !actualFields.contains(it)).collect(Collectors.joining(", "));
-            addWarningMessage(
-                    message,
-                    recordInput.getTypeName(),
-                    schema.getInputType(recordInput).getTable().getMappingName(),
-                    missingFields
-            );
-        }
-    }
-
-    private void validateNotUsingBothExternalFieldAndField() {
-        schema.getObjects().values()
-                .forEach(object -> object.getFields().stream()
-                        .filter(it -> it.isExternalField() && it.hasFieldDirective())
-                        .forEach(it -> addErrorMessage(
-                                        "Field " + it.getName() + " in type " + it.getContainerTypeName() + " cannot have both the field and externalField directives."
-                                )
-                        )
-                );
-    }
-
-    private void validateConstructTypeDirective() {
-        schema
-                .getObjects()
-                .values()
-                .forEach(object -> object
-                        .getFields()
-                        .stream()
-                        .filter(ObjectField::hasSelectConstruct)
-                        .forEach(field -> validateConstructField(field, object))
-                );
-    }
-
-    private void validateConstructField(ObjectField field, ObjectDefinition object) {
-        var targetType = schema.getObject(field.getTypeName());
-
-        if (targetType == null) {
-            addErrorMessage(CONSTRUCT_FIELD_IS_NOT_TYPE, field.formatPath(), field.getTypeName());
-            return;
-        }
-
-        if (targetType.hasTable()) {
-            addErrorMessage(CONSTRUCT_TYPE_HAS_TABLE, field.formatPath(), field.getTypeName());
-        }
-
-        var hasTable = schema.hasTableObjectForObject(object);
-        if (!hasTable) {
-            addErrorMessage(CONSTRUCT_MISSING_TABLE, field.formatPath());
-        }
-
-        if (field.hasFieldDirective()) {
-            addErrorMessage(CONSTRUCT_FIELD_HAS_ILLEGAL_DIRECTIVE, field.formatPath(), FIELD.getName());
-        }
-
-        if (field.isExternalField()) {
-            addErrorMessage(CONSTRUCT_FIELD_HAS_ILLEGAL_DIRECTIVE, field.formatPath(), EXTERNAL_FIELD.getName());
-        }
-
-        var construct = field.getSelectConstruct();
-        var tableJavaName = hasTable
-                ? schema.getPreviousTableObjectForObject(object).getTable().getMappingName()
-                : null;
-
-        for (var entry : construct.values().entrySet()) {
-            var selectionFieldName = entry.getKey();
-            var columnName = entry.getValue();
-
-            var innerField = targetType.getFields().stream()
-                    .filter(f -> f.getName().equals(selectionFieldName))
-                    .findFirst();
-
-            if (innerField.isEmpty()) {
-                addErrorMessage(CONSTRUCT_NONEXISTENT_FIELD, field.formatPath(), selectionFieldName, targetType.getName());
-            } else if (schema.isObject(innerField.get())) {
-                addErrorMessage(CONSTRUCT_CONTAINS_NESTED_TYPE, field.formatPath(), selectionFieldName);
-            }
-
-            if (hasTable && tableExists(tableJavaName) && getJavaFieldName(tableJavaName, columnName).isEmpty()) {
-                addErrorMessage(CONSTRUCT_NONEXISTENT_COLUMN, field.formatPath(), columnName, tableJavaName);
-            }
-        }
-    }
-
-    private void validateExternalField() {
-        this.schema.getObjects().values()
-                .forEach(object -> object.getFields().stream()
-                        .filter(GenerationSourceField::isExternalField)
-                        .forEach(field -> {
-                            String typeName = field.getContainerTypeName();
-                            JOOQMapping table =
-                                    Optional.ofNullable(schema.getObject(typeName).getTable())
-                                            .orElseGet(() ->
-                                                    Optional.ofNullable(schema.getPreviousTableObjectForField(field))
-                                                            .map(RecordObjectSpecification::getTable)
-                                                            .orElse(null)
-                                            );
-
-                            if (table == null) {
-                                addErrorMessage("No table found for field " + field.getName() + "in type " + typeName);
-                                return;
-                            }
-
-                            Set<String> referenceImports = GeneratorConfig.getExternalReferenceImports();
-                            List<Method> methods = referenceImports.stream()
-                                    .map(it -> getMethodFromReference(it, table.getName(), field.getName()))
-                                    .flatMap(Optional::stream)
-                                    .toList();
-
-                            if (methods.isEmpty()) {
-                                addErrorMessage("No method found for field " + field.getName() + " in type " + typeName);
-                                return;
-                            }
-
-                            if (methods.size() > 1) {
-                                addErrorMessage("Multiple methods found for field " + field.getName() + "in type " + typeName);
-                                return;
-                            }
-
-                            Method method = methods.get(0);
-
-                            if (!method.getReturnType().equals(Field.class)) {
-                                addErrorMessage("Return type of method needs to be generic type Field for field " + field.getName() + "in type " + typeName);
-                            }
-
-                            Type type = method.getGenericReturnType();
-
-                            if (type instanceof ParameterizedType paramType) {
-                                Type actualType = paramType.getActualTypeArguments()[0];
-
-                                if (!actualType.getTypeName().equals(field.getTypeClass().toString())) {
-                                    addErrorMessage("Type parameter of generic type Field in method needs to match scalar type of field " + field.getName() + "in type " + typeName);
-                                }
-                            }
-                        })
-                );
-    }
-
-    private void validateSplitQueryFieldsInJavaRecords() {
-        allFields.stream()
-                .filter(schema::hasJavaRecord)
-                .map(schema::getObject)
-                .map(AbstractObjectDefinition::getFields)
-                .flatMap(Collection::stream)
-                .filter(GenerationSourceField::createsDataFetcher)
-                .forEach(field -> {
-                    var errorMessageStart = String.format("%s in a java record has %s directive, but",
-                            field.formatPath(),
-                            SPLIT_QUERY.getName()
-                    );
-
-                    if (!schema.isObject(field) || !schema.getObject(field.getTypeName()).hasTable()) {
-                        addErrorMessage(errorMessageStart + " does not return a type with table. This is not supported.");
-                    }
-
-                    if (field.hasForwardPagination()) {
-                        addErrorMessage(errorMessageStart + " is paginated. This is not supported.");
-                    }
-                });
-    }
-
-    private void validateWrapperTypesWithPreviousTable() {
-        schema.getObjects()
-                .values()
-                .stream()
-                .filter(it -> schema.isObjectWithPreviousTableObject(it.getName()))
-                .map(AbstractObjectDefinition::getFields)
-                .flatMap(Collection::stream)
-                .filter(AbstractField::isIterableWrapped)
-                .filter(schema::isObject)
-                .filter(it -> !it.hasFieldReferences())
-                .filter(it -> !it.createsDataFetcher())
-                .filter(it -> Optional.ofNullable(schema.getObject(it).getTable())
-                        .map(t -> schema.getPreviousTableObjectForField(it).getTable().equals(t))
-                        .orElse(true)
-                )
-                .forEach(f ->
-                        addErrorMessage(
-                                "Field %s returns a list of wrapper type '%s' (a type wrapping a subset of the table fields)," +
-                                        " which is not supported. Change the field to return a single '%s' to fix.",
-                                f.formatPath(), f.getTypeName(), f.getTypeName(), f.getTypeName())
-                );
-    }
-
-    /**
-     * Validates that all fields in types with @record directive can be mapped
-     * to methods in the referenced Java record class.
-     * - Input types: validates setter methods (setXxx)
-     * - Output types: validates getter methods (getXxx)
-     */
-    private void validateJavaRecordFieldMappings() {
-        schema.getInputTypes().values().stream()
-                .filter(RecordObjectDefinition::hasJavaRecordReference)
-                .forEach(type -> validateJavaRecordMethods(type, type.getRecordReference(), true));
-
-        schema.getRecordTypes().values().stream()
-                .filter(type -> !schema.isInputType(type.getName()))
-                .filter(RecordObjectSpecification::hasJavaRecordReference)
-                .forEach(type -> validateJavaRecordMethods(
-                        (RecordObjectDefinition<?, ?>) type,
-                        type.getRecordReference(),
-                        false));
-    }
-
-    private void validateJavaRecordMethods(RecordObjectDefinition<?, ?> type, Class<?> recordClass, boolean isInput) {
-        if (recordClass == null) return;
-
-        for (var field : type.getFields()) {
-            if (field.isExplicitlyNotGenerated()) {
-                continue;
-            }
-
-            // Flatten nested types without @record/@table
-            var flattenedType = getFlattenedNestedType(field, isInput);
-            if (flattenedType.isPresent()) {
-                validateJavaRecordMethods(flattenedType.get(), recordClass, isInput);
-                continue;
-            }
-
-            if (shouldSkipFieldValidation(field, isInput)) {
-                continue;
-            }
-
-            validateFieldHasMethod(field, type, recordClass, isInput);
-        }
-    }
-
-    private Optional<RecordObjectDefinition<?, ?>> getFlattenedNestedType(GenerationField field, boolean isInput) {
-        var nested = isInput
-                ? schema.getInputType(field)
-                : schema.getObject(field.getTypeName());
-
-        if (nested != null && !nested.hasJavaRecordReference() && !nested.hasTable()) {
-            return Optional.of(nested);
-        }
-        return Optional.empty();
-    }
-
-    private boolean shouldSkipFieldValidation(GenerationField field, boolean isInput) {
-        return isInput
-                ? schema.isInputType(field)
-                : !field.createsDataFetcher() && schema.isObject(field);
-    }
-
-    private void validateFieldHasMethod(GenerationField field, RecordObjectDefinition<?, ?> type,
-                                        Class<?> recordClass, boolean isInput) {
-        if (!(field instanceof GenerationSourceField<?> sourceField)) {
-            return;
-        }
-
-        var mapping = sourceField.getJavaRecordMethodMapping(true);
-        var methodName = isInput ? mapping.asSet() : mapping.asGet();
-
-        if (ReflectionHelpers.classHasMethod(recordClass, methodName)) {
-            return;
-        }
-        var typeKind = isInput ? "input" : "type";
-        var methodType = isInput ? "setter" : "getter";
-        var suggestion = findSimilarMethods(recordClass, methodName, isInput);
-
-        var message = suggestion.isEmpty()
-                ? "Cannot map field '%s' in %s '%s' to %s in Java record '%s'. Expected method: %s"
-                : "Cannot map field '%s' in %s '%s' to %s in Java record '%s'. Expected method: %s. Did you mean: %s?";
-
-        if (suggestion.isEmpty()) {
-            addErrorMessage(message, field.getName(), typeKind, type.getName(), methodType, recordClass.getSimpleName(), methodName);
-        } else {
-            addErrorMessage(message, field.getName(), typeKind, type.getName(), methodType, recordClass.getSimpleName(), methodName, suggestion.get());
-        }
-    }
-
-    private Optional<String> findSimilarMethods(Class<?> recordClass, String expectedMethod, boolean isInput) {
-        var prefix = isInput ? "set" : "get";
-        var candidates = Arrays.stream(recordClass.getMethods())
-                .map(Method::getName)
-                .filter(name -> name.startsWith(prefix));
-
-        var similarMethods = findSimilarStringsWithDistance(expectedMethod, candidates, 12);
-        if (similarMethods.isEmpty()) {
-            return Optional.empty();
-        }
-
-        var fieldNames = similarMethods.keySet().stream()
-                .map(methodName -> uncapitalize(methodName.substring(prefix.length())))
-                .collect(Collectors.joining(", "));
-        return Optional.of(fieldNames);
-    }
-
-    /**
-     * Finds strings similar to the target using Levenshtein distance, including the distance values.
-     * @param target The string to match against
-     * @param candidates Stream of candidate strings
-     * @param maxDistance Maximum allowed distance for suggestions
-     * @return Map of similar strings to their Levenshtein distance within the minimum distance threshold (up to maxDistance)
-     */
-    private static Map<String, Integer> findSimilarStringsWithDistance(String target, Stream<String> candidates, int maxDistance) {
-        var levenshtein = new LevenshteinDistance(12);
-
-        var distances = candidates.collect(Collectors.toMap(
-                Function.identity(),
-                candidate -> levenshtein.apply(target, candidate)));
-
-        var distanceThreshold = distances.values().stream()
-                .filter(it -> it > -1 && it <= maxDistance)
-                .min(Integer::compare)
-                .orElse(0);
-
-        return distances.entrySet().stream()
-                .filter(it -> it.getValue() > -1 && it.getValue() <= distanceThreshold)
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-    }
-
-    /**
-     * Validates that paginated fields on tables without primary keys have stable ordering.
-     * Without ordering, cursor-based pagination produces unstable results.
-     * <p>
-     * For tables without primary keys, we need one of:
-     * - @defaultOrder directive (provides fallback ordering)
-     * - non-nullable @orderBy (user must always provide ordering)
-     * - both @orderBy and @defaultOrder (nullable orderBy with fallback)
-     */
-    private void validatePaginatedFieldsHaveOrdering() {
-        allFields.stream()
-                .filter(ObjectField::hasForwardPagination)
-                .filter(ObjectField::isGenerated)
-                .filter(field -> {
-                    var recordType = schema.getRecordType(field);
-                    return recordType != null && recordType.hasTable();
-                })
-                .forEach(field -> {
-                    var recordType = schema.getRecordType(field);
-                    var tableName = recordType.getTable().getMappingName();
-                    if (!tableHasPrimaryKey(tableName)) {
-                        var hasDefaultOrder = field.getDefaultOrder().isPresent();
-                        var orderByField = field.getOrderField();
-                        var hasNonNullableOrderBy = orderByField.isPresent() && orderByField.get().isNonNullable();
-
-                        if (!hasDefaultOrder && !hasNonNullableOrderBy) {
-                            if (orderByField.isPresent()) {
-                                // Has nullable @orderBy but no @defaultOrder
-                                addErrorMessage(
-                                        "Paginated field '%s' in type '%s' has nullable @%s but no @%s directive. " +
-                                                "Table '%s' has no primary key, so either make the orderBy argument non-nullable (orderBy: ...!) " +
-                                                "or add @%s to provide fallback ordering when orderBy is not provided.",
-                                        field.getName(),
-                                        field.getContainerTypeName(),
-                                        ORDER_BY.getName(),
-                                        DEFAULT_ORDER.getName(),
-                                        tableName,
-                                        DEFAULT_ORDER.getName()
-                                );
-                            } else {
-                                // No @orderBy and no @defaultOrder
-                                addErrorMessage(
-                                        "Paginated field '%s' in type '%s' requires @%s or @%s directive. " +
-                                                "Table '%s' has no primary key, so without ordering, cursor-based pagination produces unstable results.",
-                                        field.getName(),
-                                        field.getContainerTypeName(),
-                                        DEFAULT_ORDER.getName(),
-                                        ORDER_BY.getName(),
-                                        tableName
-                                );
-                            }
-                        }
-                    }
-                });
-    }
-
-    /**
-     * Validates that @defaultOrder directive is not used on fields returning interface or union types.
-     * This feature is not yet supported for interface/union queries.
-     */
-    private void validateDefaultOrderNotOnInterfaceOrUnion() {
-        allFields.stream()
-                .filter(field -> field.getDefaultOrder().isPresent())
-                .filter(field -> schema.isInterface(field) || schema.isUnion(field))
-                .forEach(field -> addErrorMessage(
-                        "@%s directive on field '%s' in type '%s' is not supported because the field returns %s type '%s'. " +
-                                "@%s is not yet supported for interface or union queries.",
-                        DEFAULT_ORDER.getName(),
-                        field.getName(),
-                        field.getContainerTypeName(),
-                        schema.isInterface(field) ? "interface" : "union",
-                        field.getTypeName(),
-                        DEFAULT_ORDER.getName()
-                ));
-    }
-
-    /**
-     * Validates that all @orderBy input types have the required structure:
-     * one enum field with @order/@index directives (the sort field) and one direction enum (ASC/DESC).
-     */
-    private void validateOrderByInputTypes() {
-        allFields.stream()
-                .map(ObjectField::getOrderField)
-                .flatMap(Optional::stream)
-                .forEach(orderField -> {
-                    var inputType = schema.getInputType(orderField);
-                    if (inputType == null) {
-                        addErrorMessage("Input type '%s' not found in schema", orderField.getTypeName());
-                        return;
-                    }
-                    var enums = inputType.getFields().stream()
-                            .filter(schema::isEnum)
-                            .map(schema::getEnum)
-                            .toList();
-                    var orderByCount = enums.stream().filter(EnumDefinition::isOrderByEnum).count();
-                    var directionCount = enums.stream().filter(EnumDefinition::isDirectionEnum).count();
-                    if (orderByCount != 1) {
-                        addErrorMessage(
-                                "Expected exactly one orderBy enum field on type '%s', but found %d. The @%s input type must contain exactly one enum field whose values have @%s directives.",
-                                orderField.getTypeName(), orderByCount, ORDER_BY.getName(), ORDER.getName());
-                    }
-                    if (directionCount != 1) {
-                        addErrorMessage(
-                                "Expected exactly one direction enum field on type '%s', but found %d. The @%s input type must contain exactly one direction enum field (e.g., with ASC/DESC values).",
-                                orderField.getTypeName(), directionCount, ORDER_BY.getName());
-                    }
-                    // Validate that all orderBy enum fields have @order or @index
-                    enums.stream()
-                            .filter(EnumDefinition::isOrderByEnum)
-                            .forEach(e -> e.getObjectDefinition().getEnumValueDefinitions().stream()
-                                    .filter(v -> !v.hasDirective(ORDER.getName()) && !v.hasDirective(INDEX.getName()))
-                                    .forEach(v -> addErrorMessage(
-                                            "Enum field '%s' of '%s' has no valid @%s directive",
-                                            v.getName(), e.getName(), ORDER.getName())));
-                });
-    }
-
-    /**
-     * Validates that neither input types nor output types contain cycles that would cause infinite recursion
-     * during code generation (or later validation).
-     */
-    private void validateNoCycles() {
-        validateInputCycles();
-        validateOutputTypeCycles();
-        throwIfErrors();
-    }
-
-    private void validateInputCycles() {
-        var alreadyExplored = new HashSet<String>();
-        schema.getInputTypes().values()
-                .forEach(input ->
-                        detectCycles(input.getName(), new LinkedHashMap<>(), alreadyExplored, "Input type cycles are not allowed."));
-    }
-
-    private void validateOutputTypeCycles() {
-        var alreadyExplored = new HashSet<String>();
-        var errorSuffix = String.format("A resolver is required to break the cycle, for example by adding @%s to (one of) the field(s).",
-                SPLIT_QUERY.getName());
-
-        schema.getObjects().values()
-                .forEach(type -> detectCycles(type.getName(), new LinkedHashMap<>(), alreadyExplored, errorSuffix));
-        schema.getInterfaces().values()
-                .forEach(type -> detectCycles(type.getName(), new LinkedHashMap<>(), alreadyExplored, errorSuffix));
-    }
-
-    /**
-     * Detects cycles in input types, and cycles in output types not broken by a resolver.
-     *
-     * @param typeName        the name of the type, interface or input currently being visited
-     * @param currentPath     ordered map of types on the current DFS path, each mapped to the field paths leading from it
-     * @param alreadyExplored types whose entire subgraph has already been checked
-     * @param errorSuffix     appended after the cycle path in the error message, e.g. a workaround suggestion
-     */
-    private void detectCycles(String typeName, LinkedHashMap<String, List<String>> currentPath, Set<String> alreadyExplored, String errorSuffix) {
-        if (alreadyExplored.contains(typeName)) return;
-
-        if (currentPath.containsKey(typeName)) {
-            addErrorMessage("Cycle detected: %s. %s", formatCyclePath(currentPath, typeName), errorSuffix);
-            return;
-        }
-
-        getNextPaths(typeName)
-                .forEach((fieldTargetType, fieldPaths) -> {
-                    currentPath.put(typeName, fieldPaths);
-                    detectCycles(fieldTargetType, currentPath, alreadyExplored, errorSuffix);
-                });
-        currentPath.remove(typeName);
-        alreadyExplored.add(typeName);
-    }
-
-    /** Returns a type's non-resolver fields, grouped by target type name. */
-    private Map<String, List<String>> getNextPaths(String typeName) {
-        if (schema.isInputType(typeName)) {
-            return schema.getInputType(typeName)
-                    .getFields().stream()
-                    .filter(GenerationTarget::isGenerated)
-                    .filter(field -> schema.isInputType(field.getTypeName()))
-                    .collect(Collectors.groupingBy(
-                            FieldSpecification::getTypeName,
-                            LinkedHashMap::new,
-                            Collectors.mapping(GenerationField::formatPath, Collectors.toList())
-                            )
-                    );
-        }
-        var fields = schema.isObject(typeName)
-                ? schema.getObject(typeName).getFields()
-                : Optional.ofNullable(schema.getInterface(typeName))
-                    .map(AbstractObjectDefinition::getFields)
-                    .orElse(List.of());
-
-        return fields.stream()
-                .filter(GenerationTarget::isGenerated)
-                .filter(field -> !field.createsDataFetcher())
-                .filter(field -> !schema.isScalar(field))
-                .flatMap(field ->
-                        resolveFieldTargets(field)
-                                .map(targetTypeName -> Map.entry(field.formatPath(), targetTypeName)))
-                .collect(Collectors.groupingBy(
-                        Map.Entry::getValue,
-                        LinkedHashMap::new,
-                        Collectors.mapping(Map.Entry::getKey, Collectors.toList())));
-    }
-
-    /** Resolves the type names a field can point to, expanding interfaces to their implementations and unions to their members. */
-    private Stream<String> resolveFieldTargets(GenerationField field) {
-        if (schema.isObject(field) || schema.isInputType(field)) {
-            return Stream.of(field.getTypeName());
-        } else if (schema.isInterface(field)) {
-            return Stream.concat(
-                    Stream.of(field.getTypeName()), // Include interface name in order to check for interface cycles
-                    schema.getImplementationsForInterface(field.getTypeName()).stream()
-                            .map(AbstractObjectDefinition::getName)
-            );
-        } else if (schema.isUnion(field)) {
-            return schema.getUnionSubTypes(field.getTypeName()).stream()
-                    .map(AbstractObjectDefinition::getName);
-        }
-        return Stream.empty();
-    }
-
-    /**
-     * Formats the cycle portion of a path as a human-readable string, e.g. {@code 'TypeA.b' -> 'TypeB.a' -> 'TypeA'}.
-     * @param fullPath Ordered map of types on the current DFS path, each mapped to the field paths leading from it
-     * @param typeNameInCycle Name of a type/interface/input type which is in the cycle.
-     * @return Path as a human-readable string
-     */
-    private static String formatCyclePath(LinkedHashMap<String, List<String>> fullPath, String typeNameInCycle) {
-        var cyclePath = new LinkedHashMap<>(fullPath);
-        for (var key : fullPath.keySet()) {
-            if (key.equals(typeNameInCycle)) break;
-            cyclePath.remove(key);
-        }
-
-        var path = cyclePath.values()
-                .stream()
-                .map(fieldPaths -> String.join("/", fieldPaths))
-                .reduce("",
-                        (accumulated, step) -> accumulated + step + " -> ");
-
-        // Append the repeated type name to close the cycle
-        return path + String.format("'%s'", cyclePath.keySet().iterator().next());
-    }
-
-    private void validateLookupArguments() {
-        allFields.forEach(field -> {
-            var lookupArguments = field.getNonReservedArguments().stream()
-                    .filter(InputField::isLookupKey)
-                    .toList();
-
-            JOOQMapping currentTable;
-            var previousTableObject = schema.getPreviousTableObjectForField(field);
-            if (schema.isObject(field.getTypeName()) && schema.getObject(field.getTypeName()).hasTable()) {
-                currentTable = schema.getObject(field.getTypeName()).getTable();
-            } else {
-                if (previousTableObject != null && previousTableObject.hasTable()) {
-                    currentTable = previousTableObject.getTable();
-                } else {
-                    currentTable = null;
-                }
-            }
-
-            lookupArguments.stream()
-                    .filter(it -> it.hasFieldReferences() || isNodeIdReferenceField(it, schema, currentTable))
-                    .forEach(it -> addErrorMessage(
-                            "Argument/input field %s has %s directive, but is a reference field. Lookup on references is not currently supported.",
-                            it.formatPath(), LOOKUP_KEY.getName()));
-
-            lookupArguments.stream()
-                    .filter(schema::isInputType)
-                    .forEach(arg -> {
-                        var referenceFields = schema.getInputType(arg).getFields().stream()
-                                .filter(it -> it.hasFieldReferences() || isNodeIdReferenceField(it, schema, currentTable))
-                                .map(GenerationSourceField::formatPath)
-                                .collect(Collectors.joining(", "));
-                        if (!referenceFields.isEmpty()) {
-                            addErrorMessage(
-                                    "Argument %s has %s directive, but contains reference field(s): %s. Lookup on references is not currently supported.",
-                                    arg.formatPath(), LOOKUP_KEY.getName(), referenceFields);
-                        }
-                    });
-        });
-    }
-
-    private void validateEntities() {
-        schema.getObjects().values().stream()
-                .filter(RecordObjectDefinition::isEntity)
-                .forEach(this::validateEntity);
-    }
-
-
-    private void validateEntity(ObjectDefinition entityType) {
-        if (!entityType.hasTable()) {
-            addErrorMessage("Entity type '%s' must map to a table using the @%s directive",
-                    entityType.getName(),
-                    TABLE.getName());
-            return;
-        }
-
-        var hasNestedKeys = entityType.getEntityKeys().keys().stream().anyMatch(key -> !key.getNestedKeys().isEmpty());
-        if (hasNestedKeys) {
-            addErrorMessage("Nested key(s) found in entity type '%s'. This is currently not supported.", entityType.getName());
-        }
-
-        for (var compositeKey: entityType.getEntityKeys().keys()) {
-            for (var key : compositeKey.getKeys()) {
-                var matchingField = entityType.getFields().stream()
-                        .filter(field -> field.getName().equals(key))
-                        .findFirst();
-                if (matchingField.isEmpty()) {
-                    var similarFields = findSimilarStringsWithDistance(key, entityType.getFields().stream().map(AbstractField::getName), 6);
-                    var suggestion = similarFields.isEmpty() ? "" : String.format(". Did you mean one of: '%s'?", String.join("', '", similarFields.keySet()));
-                    addErrorMessage("Entity Key field '%s' was not found in type '%s'%s", key, entityType.getName(), suggestion);
-                } else if (matchingField.get().hasFieldReferences() || isNodeIdReferenceField(matchingField.get(), schema)) {
-                    addErrorMessage("Entity Key field '%s' in type '%s' is a reference. This is currently not supported.",
-                            key, entityType.getName());
-                }
-            }
-
-        }
-
+        new FieldTypeValidator(schema, allFields).validate();
     }
 }
